ml-service/requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Processing
pandas==2.1.3
numpy==1.24.4
scikit-learn==1.3.2

# ML Framework (CPU version for faster build)
tensorflow-cpu==2.15.0

# Financial Data
yfinance==0.2.25

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Cache
redis==5.0.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.2

# Visualization (lightweight versions)
matplotlib==3.8.2
plotly==5.17.0

# Technical Analysis
ta==0.10.2

# HTTP Client
httpx==0.25.2

# Additional utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Scheduling
schedule==1.2.0

# Database utilities
tabulate==0.9.0